total as float in offered times

+++ app/helpers/evaluator_helper.rb
@@ -83,7 +83,7 @@ module EvaluatorHelper
   def offered_hours(project)
     offered = project.offered_hours
     if offered
-      total = project.worktimes.sum(:hours, :conditions => ['worktimes.billable']).to_i
+      total = project.worktimes.sum(:hours, :conditions => ['worktimes.billable']).to_f
       color = 'green'
       if total > offered
         color = 'red'
